Refactor Add and List methods for more OOP
