+++ src/main/java/seedu/busybreak/BusyBreak.java
@@ -1,5 +1,7 @@
 package seedu.busybreak;
 
+import seedu.busybreak.command.Add;
+import seedu.busybreak.command.List;
 import seedu.busybreak.storage.Storage;
 import seedu.busybreak.storage.Load;
 import seedu.busybreak.command.Clear;
@@ -50,11 +52,11 @@ public class BusyBreak {
             terminateProgram();
             break;
         case "list": //list out all items
-            listItems();
+            List.listItems();
             break;
         case "add": //add itinerary entry
-            ParseActivityData activityData = getParseActivityData(userInputArray);
-            addActivityDataToList(activityData);
+            Add.ParseActivityData activityData = Add.getParseActivityData(userInputArray);
+            Add.addActivityDataToList(activityData);
             break;
         case "schedule":
             setByTime();
@@ -103,7 +105,7 @@ public class BusyBreak {
                 return;
             }
 
-            ParseActivityData editedActivityData = getParseActivityData(parsedEditedInputArray);
+            Add.ParseActivityData editedActivityData = Add.getParseActivityData(parsedEditedInputArray);
             assert editedActivityData != null : "Parsed activity data cannot be null";
 
             list.set(index, new Activity(editedActivityData.date(), editedActivityData.time(),
@@ -170,119 +172,6 @@ public class BusyBreak {
         }
     }
 
-    private static void addActivityDataToList(ParseActivityData activityData) {
-        if (activityData == null) {
-            return;
-        }
-
-        list.add(new Activity(activityData.date(), activityData.time(),
-                activityData.description(), activityData.cost()));
-        System.out.println(LINE);
-        System.out.print("Added Activity to Itinerary: ");
-        System.out.print("Date: " + activityData.date() + " | ");
-        System.out.print("Time: " + activityData.time() + " | ");
-        System.out.print("Description: " + activityData.description() + " | ");
-        System.out.println("Cost: $" + activityData.cost());
-        System.out.println(LINE);
-
-        storage.saveActivities();
-    }
-
-    private static ParseActivityData getParseActivityData(String[] userInputArray) {
-
-        try {
-            //truncate the command from userInput
-            String[] parsedUserInputArray = Arrays.copyOfRange(userInputArray, 1, userInputArray.length);
-            String parsedUserInput = String.join(" ", parsedUserInputArray).trim();
-
-            assert !parsedUserInput.isEmpty() : "User input cannot be empty.";
-
-            String[] parseDate = getParsedDate(parsedUserInput);
-            String[] parseTime = getParsedTime(parseDate);
-            String[] parseDescription = getParsedDescription(parseTime);
-            String[] parseCost = getParsedCost(parseDescription);
-
-            String date = parseTime[0].trim();
-            String time = parseDescription[0].trim();
-            String description = parseCost[0].trim();
-            String cost = parseCost[1].trim();
-            checkEmptyFields(date, time, description, cost);
-
-            assert !date.isEmpty() : "Date cannot be empty.";
-            assert !time.isEmpty() : "Time cannot be empty.";
-            assert !description.isEmpty() : "Description cannot be empty.";
-            assert !cost.isEmpty() : "Cost cannot be empty.";
-
-            return new ParseActivityData(date, time, description, cost);
-        } catch (IllegalArgumentException | AssertionError e) {
-            System.out.println(LINE);
-            System.out.println(e.getMessage());
-            System.out.println(LINE);
-            return null;
-        }
-    }
-
-    private static void checkEmptyFields(String date, String time, String description, String cost) {
-        if (date.isEmpty() || time.isEmpty() || description.isEmpty() || cost.isEmpty()) {
-            throw new IllegalArgumentException("User input is invalid! One or more fields :" +
-                    " date, time, description or cost is empty.");
-        }
-    }
-
-    private static String[] getParsedCost(String[] parseDescription) {
-        String[] parseCost = parseDescription[1].split("c/", 2);
-        if (parseCost.length != 2) {
-            throw new IllegalArgumentException("Missing cost! Set one with c/");
-        }
-        return parseCost;
-    }
-
-    private static String[] getParsedDescription(String[] parseTime) {
-        String[] parseDescription = parseTime[1].split("desc/", 2);
-        if (parseDescription.length != 2) {
-            throw new IllegalArgumentException("Missing description! Set one with desc/");
-        }
-        return parseDescription;
-    }
-
-    private static String[] getParsedTime(String[] parseDate) {
-        String[] parseTime = parseDate[1].split("t/", 2);
-        if (parseTime.length != 2) {
-            throw new IllegalArgumentException("Missing time! Set one with t/");
-        }
-        return parseTime;
-    }
-
-    private static String[] getParsedDate(String parsedUserInput) {
-        String[] parseDate = parsedUserInput.split("d/", 2);
-        if (parseDate.length != 2) {
-            throw new IllegalArgumentException("Missing date! Set one with d/");
-        }
-        return parseDate;
-    }
-
-    private record ParseActivityData(String date, String time, String description, String cost) {
-    }
-
-    private static void listItems() {
-
-        if (list.isEmpty()) {
-            System.out.println(LINE);
-            System.out.println("Itinerary is Empty!");
-            System.out.println(LINE);
-            return;
-        }
-        System.out.println(LINE);
-        for (int index = 0; index < list.size(); index++) {
-            System.out.println((index + 1) + ". ");
-            System.out.println("Date: " + list.get(index).getDate());
-            System.out.println("Time: " + list.get(index).getTime());
-            System.out.println("Description: " + list.get(index).getDescription());
-            System.out.println("Cost: $" + list.get(index).getCost());
-            System.out.println(LINE);
-        }
-    }
-
     private static void setByTime() {
         if (list.isEmpty()) {
             System.out.println(LINE);
@@ -295,7 +184,7 @@ public class BusyBreak {
 
         System.out.println(LINE);
         System.out.println("Your Activities are sorted by time now!");
-        listItems();
+        List.listItems();
 
         storage.saveActivities();
     }
